requirement.txt
openai
langchain
streamlit
python-dotenv
PyPDF2
huggingface_hub
transformers
accelerate
bitsandbytes
langchain_openai
langchain-community

-e .